Split into two sumcheck instances, OutputSumcheck and ValFinalSumcheck
